pyproject.toml
[build-system]
requires = [
    "scikit-build-core>=0.8",
    "swig>=4.0",
    "numpy>=1.20",
]
build-backend = "scikit_build_core.build"

[project]
name = "occt-rt"
version = "1.0.0"
description = "High-performance BVH raytracer for OpenCASCADE"
readme = "README.md"
license = {text = "LGPL-2.1-or-later"}
authors = [
    {name = "Andrea Pozzetti", email = "andrea@example.com"}
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: C++",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Visualization",
]
keywords = ["cad", "opencascade", "raytracing", "bvh", "geometry"]

[project.optional-dependencies]
pythonocc = ["pythonocc-core>=7.8"]
embree = []  # Embree is a C++ dependency, not pip-installable
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.urls]
Homepage = "https://github.com/PozzettiAndrea/occt-rt-python"
Repository = "https://github.com/PozzettiAndrea/occt-rt-python.git"
Issues = "https://github.com/PozzettiAndrea/occt-rt-python/issues"

[tool.scikit-build]
cmake.minimum-version = "3.18"
cmake.build-type = "Release"
wheel.packages = ["occt_rt"]
build-dir = "build/{wheel_tag}"

[tool.scikit-build.cmake.define]
CMAKE_BUILD_TYPE = "Release"
